# docs/swagger.yaml
basePath: /
definitions:
...
  common.PaginatedResponse-dto_UserResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.UserResponse'
        type: array
      pagination:
        $ref: '#/definitions/common.PaginationInfo'
    type: object
  common.PaginationInfo:
    properties:
      current_page:
        type: integer
      last_page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
    type: object
...
  dto.ItemResponse:
    properties:
      code:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  dto.ProfileSummary:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  dto.UserCreateRequest:
    properties:
      email:
        example: cesar@tudominio.com
        maxLength: 255
        type: string
      name:
        example: César Pérez
        maxLength: 255
        minLength: 2
        type: string
      password:
        example: '********'
        maxLength: 100
        minLength: 8
        type: string
      phone:
        description: Metadata (opcional)
        example: "+56912345678"
        maxLength: 20
        type: string
      profile_id:
        example: 1
        type: integer
      state:
        enum:
        - 0
        - 1
        example: 1
        type: integer
    required:
    - email
    - name
    - password
    type: object
  dto.UserModuleResponse:
    properties:
      items:
        description: ← Items con code incluido
        items:
          $ref: '#/definitions/dto.ItemResponse'
        type: array
      name:
        type: string
      slug:
        type: string
    type: object
  dto.UserResponse:
    properties:
      date:
        type: string
      email:
        type: string
      id:
        type: integer
      modules:
        description: ← Usa el tipo específico
        items:
          $ref: '#/definitions/dto.UserModuleResponse'
        type: array
      name:
        type: string
      phone:
        type: string
      profile:
        $ref: '#/definitions/dto.ProfileSummary'
      profile_id:
        type: integer
      state:
        type: integer
      time:
        type: string
    type: object
  dto.UserUpdateRequest:
    properties:
      email:
        example: cesar@tudominio.com
        maxLength: 255
        type: string
      name:
        example: César Pérez
        maxLength: 255
        minLength: 2
        type: string
      password:
        example: '********'
        maxLength: 100
        minLength: 8
        type: string
      phone:
        example: "+56912345678"
        maxLength: 20
        type: string
      profile_id:
        example: 1
        type: integer
      state:
        enum:
        - 0
        - 1
        example: 1
        type: integer
    type: object
host: localhost:8082
info:
  contact: {}
  description: API Backend para GOLANG DEMO.
  title: Golang DEMO
  version: "1.0"
paths:
...
  /users:
    get:
      description: Retorna usuarios paginados con búsqueda por nombre/email, filtros
        por estado y perfil, y ordenamiento flexible
      parameters:
      - default: 1
        description: Página
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Registros por página
        in: query
        maximum: 100
        minimum: 1
        name: per_page
        type: integer
      - description: Término de búsqueda
        in: query
        name: search
        type: string
      - default: name
        description: Campo a buscar
        enum:
        - name
        - email
        in: query
        name: field
        type: string
      - description: Filtrar por estado
        enum:
        - 0
        - 1
        in: query
        name: state
        type: integer
      - description: Filtrar por perfil
        in: query
        name: profile_id
        type: integer
      - default: id
        description: Campo para ordenar
        enum:
        - id
        - name
        - email
        - created_at
        in: query
        name: sort_by
        type: string
      - default: desc
        description: Dirección de ordenamiento
        enum:
        - asc
        - desc
        in: query
        name: sort_dir
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.PaginatedResponse-dto_UserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Listar usuarios con paginación y filtros avanzados
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Crea un usuario con validación de email único, password hasheado
        con bcrypt, y metadata opcional
      parameters:
      - description: Datos del usuario
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UserCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Crear nuevo usuario
      tags:
      - Users
  /users/{id}:
    delete:
      description: Elimina un usuario y su metadata asociada de forma segura
      parameters:
      - description: ID del usuario
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Eliminar usuario
      tags:
      - Users
    get:
      description: Retorna un usuario específico con su metadata y perfil asociado
      parameters:
      - description: ID del usuario
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Obtener usuario por ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Actualiza datos del usuario, metadata y password (opcional) con
        validaciones robustas
      parameters:
      - description: ID del usuario
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: Nuevos datos del usuario
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Actualizar usuario
      tags:
      - Users
  /users/me:
    get:
      description: Retorna los datos del usuario autenticado (requiere middleware
        de auth)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Obtener usuario autenticado
      tags:
      - Users
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"